Changes imported from Abseil "staging" branch:
  - db061dd2b795837e6154be0991077ca5a03ec471 Release the async-signal safe demangler as an internal im... by Derek Mauro <dmauro@google.com>
  - 310440eb33d946df79d26bf1edf795f6a0b466b9 Use static_cast to constrain nanoseconds to int range, av... by Abseil Team <absl-team@google.com>

GitOrigin-RevId: db061dd2b795837e6154be0991077ca5a03ec471

+++ absl/base/internal/spinlock_wait.cc
@@ -72,7 +72,7 @@ int SpinLockSuggestedDelayNS(int loop) {
   // Mean is exponential in loop for first 32 iterations, then 8ms.
   // The futex path multiplies this by 16, since we expect explicit wakeups
   // almost always on that path.
-  return r >> (44 - (loop >> 3));
+  return static_cast<int>(r >> (44 - (loop >> 3)));
 }
 
 }  // namespace base_internal
